requirements.txt
click==8.1.7
openai==1.3.5
playwright==1.40.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
requests==2.31.0
httpx>=0.24.1,<1.0.0
